Criação do controller

+++ ecommerce-shopping-api/src/main/java/com/rendersonjunior/ecommerceshoppingapi/model/Shop.java
@@ -1,21 +1,13 @@
 package com.rendersonjunior.ecommerceshoppingapi.model;
 
 
-import jakarta.persistence.CollectionTable;
-import jakarta.persistence.Column;
-import jakarta.persistence.ElementCollection;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -36,7 +28,7 @@ public class Shop {
     private String userIdentifier;
 
     @Column(name = "TOTAL")
-    private Double total;
+    private BigDecimal total;
 
     @Column(name = "DATE")
     private LocalDateTime date;
