+++ ecommerce-shopping-api/src/main/java/com/rendersonjunior/ecommerceshoppingapi/repository/ShopRepository.java
@@ -4,6 +4,7 @@ import com.rendersonjunior.ecommerceshoppingapi.model.Shop;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -12,7 +13,7 @@ public interface ShopRepository extends JpaRepository<Shop, Long> {
 
     List<Shop> findAllByUserIdentifier(String userIdentifier);
 
-    List<Shop> findAllByTotalGreaterThan(Double total);
+    List<Shop> findAllByTotalGreaterThan(BigDecimal total);
 
     List<Shop> findAllByDateGreaterThan(LocalDateTime date);
